+++ src/main/java/me/villagerunknown/coinbanks/block/ToughJarCoinBankBlock.java
@@ -1,0 +1,74 @@
+package me.villagerunknown.coinbanks.block;
+
+import com.google.common.collect.ImmutableMap;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.particle.ParticleTypes;
+import net.minecraft.particle.SimpleParticleType;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.util.Hand;
+import net.minecraft.util.ItemActionResult;
+import net.minecraft.util.hit.BlockHitResult;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+import net.minecraft.world.World;
+
+import java.util.HashMap;
+import java.util.function.Function;
+
+public class ToughJarCoinBankBlock extends CoinBankBlock {
+	
+	protected static final VoxelShape SHAPE = Block.createCuboidShape(5.0, 0.0, 5.0, 11.0, 8.0, 11.0);
+	
+	public ToughJarCoinBankBlock(Settings settings) {
+		super(
+				settings
+						.nonOpaque()
+						.hardness( Blocks.OBSIDIAN.getHardness() )
+						.resistance( Blocks.OBSIDIAN.getBlastResistance() )
+		);
+	}
+	
+	@Override
+	protected ItemActionResult onUseWithItem(ItemStack stack, BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+		ItemActionResult result = super.onUseWithItem(stack, state, world, pos, player, hand, hit);
+		
+		if( result.equals( ItemActionResult.CONSUME ) ) {
+			MinecraftServer server = world.getServer();
+			if( null != server ) {
+				ServerWorld serverWorld = server.getWorld(world.getRegistryKey());
+				if( null != serverWorld ) {
+					SimpleParticleType particleType = ParticleTypes.DUST_PLUME;
+					
+					if( state.get(WATERLOGGED) ) {
+						particleType = ParticleTypes.BUBBLE;
+					} // if
+					
+					serverWorld.spawnParticles(particleType, (double) pos.getX() + (double) 0.5F, (double) pos.getY() + 0.6, (double) pos.getZ() + (double) 0.5F, 7, (double) 0.0F, (double) 0.0F, (double) 0.0F, (double) 0.0F);
+				} // if
+			} // if
+		} // if
+		
+		return result;
+	}
+	
+	@Override
+	protected ImmutableMap<BlockState, VoxelShape> getShapesForStates(Function<BlockState, VoxelShape> stateToShape) {
+		HashMap<BlockState, VoxelShape> shapes = new HashMap<>();
+		
+		shapes.put( stateManager.getDefaultState(), SHAPE );
+		
+		return ImmutableMap.<BlockState, VoxelShape>builder().putAll( shapes ).build();
+	}
+	
+	protected VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+		return SHAPE;
+	}
+	
+}